Composing Components
Re-using the counter component from the counter-app. Created a new 'counters' component to render mulitple 'counter' componenets inside(making a tree of components)
Updated the index.js to use counters component.
Added a counters array to the state to uniquely identify each counter.
Used 'map' method to map counters array to the Counter component.
If we change the value in the counters array it'll not be passed to the counter component, we'll fix that in the next section.