Passing Data to Components -> Props
Every react component has a property called props. It's basically a plain JS object that includes all the attributes we set in the counter component. So the attributes we set on the counter 'value' and 'selected' will be the properties of that props object. Key will not be included as it's a special attribute for uniquely identifying elements.
Updated the counter component state to replace count with 'value' and used the VLAUE which was sent as a props to intialize the counter component value.