Props vs State
Props includes data that we give to a component (Input to the component), whereas state includes data that is local/private to that component so other components can't acess that state. It's completely internal to that component.
Sometime a component might not have a state it may get all the data that it needs via props.
Props is read-only. We can't change input to the component inside of that component.
If we need to modify the input during the lifecycle of a component then we need to get that input and put it in the state