Raising and Handling Events -> Delete button for each counter
We want to add delete button for each counter.
Rule of thumb in REACT app is 'The component that own a piece of state should be the one modifying it'
So, modified our counter component to raise an event 'onDelete' and counters component will handle this event using 'handleDelete'.
Basically, added a new method in our counters component and passed a reference to that method via props to the counter component