Updating the state
Updated the handleDelete method to be an arrow function for binding 'this'.
Instead of passing the properties separately, passed the counter object via props to avoid passing each property separately via props as the whole point of object is to group data.
We won't be able to access key attribute via props as key is used internally.