Single source of Truth
Each of the component has their own local state, clicking the reset button updates the value of counters but it isn't displayed on the view as both values are disconnected becuase we don't have a single source of truth. We need to remove the local state in our counter component and have a single source of truth.