Multiple Components in Sync
Brought back the App component in index. Updating the component tree to have App as a root with navbar and counters as children. Added NavBar component.
We passed Counters state to the Counter via props as we had parent-child relationship b/w the two. But there is no parent/child relationship b/w Navbar and Counters so. we need to lift the state up, so we'll lift the state of Counters to App, now both Counters and NavBar have same parent and we can pass the state via props.