Stateless functional component
NavBar component only have render mehtod, we don't have any state, any event handlers or helper methods and are getting all the data via props. In this situation we can use Stateless functional component. It simply returns the react element. We've to add props as a parameter to stateless functional component as this.props only works with class components.