Lifecycle Hooks -> Mount, Update, UnMount
Our components go through various phases during their lifecycle.
The first phase is the mounting phase, and this is when an instance of a component is created and inserted into DOM. There are few special methods that we can add to our components and react will automatically call these methods. we refer to these methods as lifecycle hooks.  So, they allow us to hook into certain moments during the lifecylce of a component and do something.
In the mounting phase we have 3 lifecycle hooks, 1-constructor, 2-render and 3-componentDidMount. React will call these methods in order.
Second lifecycle phase is update, and it happens when the state or the props of a component get changed. In this phase we have two lifecycle hooks, render and componentDidUpdate.
Last Phase is the UnMount phase, this is when a component is removed from DOM such as when we delete a counter.