Mounting Phase
Added constructor in the App component. Common usecase is to set the state based on the props received using this.state. We can't use this.setState in the constructor, we'll get an error if we do that becuase this.setState can only be called when a component is rendered and placed in DOM. So, you can set the state directly in the constructor.
In order to use this.props inside the constructor you'll to pass props as a parameter to the constructor.
Next lifecycle hook is componentDidMount, it is called after the component is rendered into the DOM and is perfect place to make ajax calls to get data from the server.
Third lifecycle hook is render. When a component is rendered all it's children are rendered recursively.
You can't use lifecycle hooks in stateless functional components. If you need to use lifecycle hooks you'll have to use calss components