Unmounting Phase
We only have one lifecycle hook, i-e componentWillUnmount, this method is called just before the component is removed from the DOM.
As a result of deleting the counter, state of App component is changed, so our entire component tree is re-rendered.
This gives us an opportunity to do any kind of clean up, like timers or listeners otherwise you'll end up with memory leaks.