Excersise Like Component
Added a folder named common. We'll add all the reusable components in this folder that are not specific to the app like pagination, tables and forms.
Converted the Like class component to stateless functional component as we don't have any state or helper methods.
Installed FontAwesome 5 using react-fontawesome pkg. Implemented the handleLike handler in the movie component.