Pagination - Type Checking with PropTypes
Installed prop-types@15.6.2 to do type checking so that the consumer passes the required props with the right type to avoid any bugs in the app.
It's good practice to use propTypes to catch bugs related to type checking. It also give documentation to your components, you can simply look at PropTypes to look at all the props, their types and whether they're required or optional.