Filtering - Adding All Genres
To add All Genres, we've only used the name property, we don't need an _id as this item is not coming from a database.
Updated the filtering algo as All Genres don't have an _id.
And whenever we change the genre, we should alos reset the currentPage to 1 to avoid bugs in the app.