Sorting - Implementing Sorting
Intitalized the sortColumn in the state. To implement sorting on the client we need lodash.
We should do sorting after filtering the data. Used orderBy method. First argument is input Array, second argument is an array of property names, third argument is an array of sort order.
To implement the descending sort, updated the sort algo so that, if the path is same update the order otherwise update the path and set the path to the new path and set the order to be asc.