Adding Routing
In React there is no concept of routing, react is only responsible for rendering a view, so to add routing we need to install a library called react-router-dom@4.3.1.
Imported the BrowserRouter in the index.js.
Wrapped app component with BrowserRouter.
BrowserRouter wraps history object in browsers and passes it down to the component tree.
Next we need to register our routes(using Route component), which basically means we need to tell react what component should be rendered based on the given url.
Route is a component, it has two attributes which will be passed as props (path and component)