Link
At the moment, moving to a link causes http requests to the server reloading the entire page. However, in Single Page Application we should only update what we've in the content area.
To turn our app into SPA, we need to replace all anchor elements in NavBar with Link component from react-router-dom
Link component doesn't have `href` attribute instead it has `to` attribute. Link will update the URL and once URL is updated one of the routes will match depending on the current URL and we'll get a new component in the DOM