Route Props
Esentially Route component is a wrapper around the component we pass to the Route. If the path matches, it'll render the component and will automatically inject three props (history, location, match)
We use history to work with the history Object in the browser and with that we can send the user to a different page.
Location refers to where the app is now.
match contains info about how this URL match the path we set in the Route.