Passing Props
When using Route, if you need to pass additional props to the component, instead of using the component attribute you should use render attribute in Route component
So, updated the Route component to use render attribute. When using render attribute we need to pass props as a paraemter to the arrow function, and then use a special syntax in JSX to destructure the props argument.