Route Parameters
We've defined a new route for our product details page. To define a parameter we should prefix it with a colon, we can also define multiple parameters.
We've put this new route as a first route as this route is more specific like/products/1.
To read the Route parameters passed to a component we use match object in props.
Replaced the anchor in the products component with Link component to avoid page refresh on loading a product details component.