Redirects
When we go to an invalid route, we're redirected to the home component but we want to redirect the user to a Not Found page.
Imported Redirect component from react-router-dom. Added exact attribute to the Home Route and a Redirect.
Sometime you want to move resources from one URL to another url. you set two attributes from and to.