Programmatic Navigation
When a user submit a form(like adding a product) we want to take the user back to the Products page this is called programmatic navigation.
History object has a bunch of useful methods for navigation such as goBack, goForward as well as push and replace.
Difference b/w push and replace is that push method will add the new address in the browser history so we can click the back button and go back where we were before.
Replace basically replaces the current address so we won't have history.
A usecase for replace method is to use it with login pages, so when the user logins and goes to a page you don't want to take the user back to login page if they click the back button.