Vidly - Adding Routes
Added App component to index and wrapped it with BrowserRouter.
Defined Routes in the App component. Used switch to render the first child in Routes that matches the location, so the other Routes will be ignored.
Used Redirect to redirect the homepage to the movies component and to the NotFound component