Refs - Access DOM element
When submitting the form we need values of input fields. In order to get access to a DOM element(input in our case) we need to give it a reference.
You should not directly access the Document object to get access to input field, use React ref instead.
As a rule of thumb, you should minizme the use of refs, use it only when you know whar you're doing.
Used ref to set focus on a given input field.