Controlled Elements
When building forms, quite often our forms have a state that is intialized based on what we get from a server. So, we've added a state property to the LoginForm
To sync the username field with the state, added a value attribute to the username field.
When the user type soemthing in the username nothing shows to fix this we've to listen for the onChange event and it's handler updates the state which results in re-rendering the form.