Handling Multiple Inputs
To handle multiple inputs we've updated the handleChange to use bracket notation to access the value of the currentTarget using the name attribute set on the input fields.
As we're only access the currentTarget from the event e so we use object destructuring to get currentTarget as input in the parameter.