Validation
To handle the errors we've added an object errors in the state of LoginForm. Initially we set it to an empty object.
We validate the form when the form is submitted. Added a very simple validate method which returns an Errors object to update the state. And if there are errors we immediately return without calling the server/database.