Code review
We're going to review the code to look for possibilities to extract our code to a reusable component.
Every form should have errors property in the State object. Validate method is also reuable piece of code except we'll have to come up with a convention and assume that all in our forms in the state object we'll have a data property that will represent data behind that form. Here data should be set to account object, in other form data can be a different object. Same is true for validateProperty. handleChange and onSubmit are tricky. So over the next few lectures we'll do refactoring.