Implemented Search
Added a new controlled component SearchBox, encapsulated the input field inside a search box, so we would have a simpler interface to work with. Intialized the searchQuery and selectedGenre in the state.
Updated the handleGenreSelect to update the searchQuery to be emty string instead of null as SearchBox is a controlled component and we should not use null/undefined with controlled elements.
Updated the filtering logic in the movie component.