Http Clients - Axios
We need to send http requests to the backend to get/save data. As react is just a lightweight library for building UI's all it cares about is rendering the view and manging the DOM, it makes sure that what we've in browser DOM is in sync with the state of your application.
In React you can use your preffered library to send http requests.
Your options for making Http requests are
Fetch API, jQuery Ajax and Axios.
Installed axios@0.18