Getting Data
Right place to call the server to get some data is in componentDidMount lifecycle.
Now we call get method to send an http request and get some data. First argument is the URL, this method returns a promise.
A promise is an object that holds the resolve of an asynchronous operation. It is an operation which is going to complete in the future. Intially the promise is in pending state, when the operation completes the state will change to resolved or rejected.
You can use traditional promise using .then or can use async await.
When you use await keyword in a function, you should decorate that function with the async keyword.