Lifecycle of a Request
Common http methods are GET, POST, PUT and DELETE. But we've an http OPTIONS request it's because our app is hosted on localhost and backend is hosted somewhere else(JSON placeholder), so for security reasons whenever an app sends an ajax request with different domain browser always sends an OPTIONS request.
Second request is POST, it's status is 201 which means resource is created. Under request payload, you can see an object we sent to the server. In the response tab, you can see response that JSON placeholder sent to us.
So, Axios wraps this object in a response object, that res object has a data property that we picked using object destructuring syntax and this is the lifecycle of a request.