Updating Data - Axios Patch and Put
Difference b/w patch and put is that, in case of patch we only send a property that needs to be updated, and in case of put we send entire object to the server.
In order to update an object/resource we send an http request to a given endpoint with id of that resource, server gets that id and the object we put in the body of the request and then put that object in the body of the response.