Optimistic vs Pessimistic Updates
We're doing pessimistic updates at the moment, there is a delay in updating the UI because we're calling the server first and then updating the UI, more optimistic approach would be to update the UI first, then call the server and if any error occurs, revert the state to original. In optimistic apporach we keep a reference to the original state, then we update the UI before calling the server and then wrap the server call in try-catch block.