Expected vs Unexpected errors
Expected(404: not found, 400: bad request[when you submit a form with invalid data]) - CLIENT ERRORS.
In case of 400 range errors, Display a specific error message.
Unexpected(network down, server down, db down, bug) - Log them - Display a generic and friendly error message.
Exception object have two properties request and response.
response is set if we successfuly get a response from a server, if the network is down or server is down you'll not get a response back so it'll be null.
request will be set if we successfully submit a request to the server otherwise it'll be null.