Handling Unexpected Errors Globally
We want to display generic error for our CRUD operations and we use Axios intercepter. In Axios we can intercept our requests and responses and if we've a response with an unexpected error, we can handle that error in one place.
We can intercept all our requests and responses.
use method takes two parameters both these are functions, first will be called if the response is successful, second will be called if response includes an error.

Error function will be excuted everytime a response has an error. To pass control to the  our catch block we need to return a rejected promise.