Extracting a Reusable Http Service - Removed comments and extra code from the App component
Handling errors globally has polutted our App module.
For another react app we'll have to copy/paste this implementation which isn't portable
It'll be better if we've a separate module like http service module and all this code put into that module and we can copy/paste that file in any new project.
We've hidden axios behind our httpService module, on the future if we want to update the library we just need to change our httpService module.
In our App module we don't care about the library we're using for http requests.