Displaying Toast Notifications - react-toastify@4.1
Installed react-toastify@4.1
In the App imported a component ToastContainer from this module. Also imported it's CSS library.
Added ToastContainer in the React.Fragment.
In httpService replaced alert with toast.error
Remember toast is a function, and as functions are objects, so we can use this toast as an object and call error method or we can use it as a function.
If you want to display info messages to the users, you can simply call the toast function otherwise can call the mehtods like error, info, success etc.