Extracting a Logger Service  - Refactor (logService created)
We've polluted our index module with details about setting up error logging service, index is like a manager and the details of logging service shouldn't be here.
Second issue is if we want to replace the Sentry with another service we'll have to search the entire project to modify the reference of sentry and the replaced library might not have the captureExpection method or it requires different arguments etc so we'll have to make a lot of changes to replace sentry. To avoid this, we'll extract all the logic of working with Sentry.io to a separate module.