Connect Movies Page to the Backend - Adding http and logService
Reused both these services from http-app folder. Created genreService with the same interface as fakeGenreService as we want to minimize the impact of changes.
Installed axios@0.18.0 and react-toastiy@4.1.0 as these are dependencies of httpService. Added ToastContainer to App.js
We don't want to use Sentry for the logging so we'll comment the sentry code in logservice and will just log to the console for now.